Created a new TTI example. In doing so, cleaned up the older acoustic example. The class AcousticWave2D is now redundant within this project (the outer project still needs it)
