+++ examples/fwi_operators.py
@@ -1,9 +1,135 @@
 from devito.operators import *
 from sympy import Eq
-from devito.interfaces import TimeData, DenseData
+from devito.interfaces import TimeData, DenseData, PointData
 from sympy import Function, symbols, as_finite_diff, Wild, IndexedBase
 from sympy.abc import x, y, t, z
-from sympy import solve
+from sympy import solve, Matrix
+
+
+class SourceLike(PointData):
+    """Defines the behaviour of sources and receivers.
+    """
+    def __init__(self, name, npoint, nt, dt, h, data, ndim, dtype, nbpml):
+        self.orig_data = data
+        self.dt = dt
+        self.h = h
+        self.ndim = ndim
+        self.nbpml = nbpml
+        super(SourceLike, self).__init__(name, npoint, nt, dtype)
+        x1, y1, z1, x2, y2, z2 = symbols('x1, y1, z1, x2, y2, z2')
+        if ndim == 2:
+            A = Matrix([[1, x1, z1, x1*z1],
+                        [1, x1, z2, x1*z2],
+                        [1, x2, z1, x2*z1],
+                        [1, x2, z2, x2*z2]])
+            self.increments = (0, 0), (0, 1), (1, 0), (1, 1)
+            self.rs = symbols('rx, rz')
+            rx, rz = self.rs
+            p = Matrix([[1],
+                        [rx],
+                        [rz],
+                        [rx*rz]])
+        else:
+            A = Matrix([[1, x1, y1, z1, x1*y1, x1*z1, y1*z1, x1*y1*z1],
+                        [1, x1, y2, z1, x1*y2, x1*z1, y2*z1, x1*y2*z1],
+                        [1, x2, y1, z1, x2*y1, x2*z1, y2*z1, x2*y1*z1],
+                        [1, x1, y1, z2, x1*y1, x1*z2, y1*z2, x1*y1*z2],
+                        [1, x2, y2, z1, x2*y2, x2*z1, y2*z1, x2*y2*z1],
+                        [1, x1, y2, z2, x1*y2, x1*z2, y2*z2, x1*y2*z2],
+                        [1, x2, y1, z2, x2*y1, x2*z2, y1*z2, x2*y1*z2],
+                        [1, x2, y2, z2, x2*y2, x2*z2, y2*z2, x2*y2*z2]])
+            self.increments = (0, 0, 0), (0, 1, 0), (1, 0, 0), (0, 0, 1), (1, 1, 0), (0, 1, 1), (1, 0, 1), (1, 1, 1)
+            self.rs = symbols('rx, ry, rz')
+            rx, ry, rz = self.rs
+            p = Matrix([[1],
+                        [rx],
+                        [ry],
+                        [rz],
+                        [rx*ry],
+                        [rx*rz],
+                        [ry*rz],
+                        [rx*ry*rz]])
+
+        # Map to reference cell
+        reference_cell = [(x1, 0),
+                          (y1, 0),
+                          (z1, 0),
+                          (x2, self.h),
+                          (y2, self.h),
+                          (z2, self.h)]
+        A = A.subs(reference_cell)
+        self.bs = A.inv().T.dot(p)
+
+    def point2grid(self, pt_coords):
+        # In: s - Magnitude of the source
+        #     x, z - Position of the source
+        # Returns: (i, k) - Grid coordinate at top left of grid cell.
+        #          (s11, s12, s21, s22) - source values at coordinates
+        #          (i, k), (i, k+1), (i+1, k), (i+1, k+1)
+        if self.ndim == 2:
+            rx, rz = self.rs
+        else:
+            rx, ry, rz = self.rs
+        x, y, z = pt_coords
+        i = int(x/self.h)
+        k = int(z/self.h)
+        coords = (i + self.nbpml, k + self.nbpml)
+        subs = []
+        x = x - i*self.h
+        subs.append((rx, x))
+        if self.ndim == 3:
+            j = int(y/self.h)
+            y = y - j*self.h
+            subs.append((ry, y))
+            coords = (i + self.nbpml, j + self.nbpml, k + self.nbpml)
+        z = z - k*self.h
+        subs.append((rz, z))
+        s = [b.subs(subs).evalf() for b in self.bs]
+        return coords, tuple(s)
+
+    # Interpolate onto receiver point.
+    def grid2point(self, u, pt_coords):
+        if self.ndim == 2:
+            rx, rz = self.rs
+        else:
+            rx, ry, rz = self.rs
+        x, y, z = pt_coords
+        i = int(x/self.h)
+        k = int(z/self.h)
+
+        x = x - i*self.h
+        z = z - k*self.h
+
+        subs = []
+        subs.append((rx, x))
+
+        if self.ndim == 3:
+            j = int(y/self.h)
+            y = y - j*self.h
+            subs.append((ry, y))
+        subs.append((rz, z))
+        if self.ndim == 2:
+            return sum([b.subs(subs) * u[t, i+inc[0]+self.nbpml, k+inc[1]+self.nbpml] for inc, b in zip(self.increments, self.bs)])
+        else:
+            return sum([b.subs(subs) * u[t, i+inc[0]+self.nbpml, j+inc[1]+self.nbpml, k+inc[2]+self.nbpml] for inc, b in zip(self.increments, self.bs)])
+
+    def read(self, u):
+        eqs = []
+        for i in range(self.npoints):
+            eqs.append(Eq(self[t, i], self.grid2point(u, self.orig_data[i, :])))
+        return eqs
+
+    def add(self, m, u):
+        assignments = []
+        dt = self.dt
+        for j in range(self.npoints):
+            add = self.point2grid(self.orig_data[j, :])
+            coords = add[0]
+            s = add[1]
+            assignments += [Eq(u[tuple([t] + [coords[i] + inc[i] for i in range(self.ndim)])],
+                               u[tuple([t] + [coords[i] + inc[i] for i in range(self.ndim)])] + self[t, j]*dt*dt/m[coords]*w) for w, inc in zip(s, self.increments)]
+        filtered = [x for x in assignments if isinstance(x, Eq)]
+        return filtered
 
 
 class FWIOperator(Operator):
