+++ examples/tti_operators.py
@@ -1,11 +1,21 @@
 from devito.operators import *
 from sympy import Eq
 from devito.interfaces import TimeData, DenseData
-from sympy import Function, symbols, as_finite_diff, Wild, IndexedBase
+from sympy import Function, symbols, as_finite_diff, Wild
 from sympy.abc import x, y, t, z
 from sympy import solve
 from sympy import *
 from sympy.abc import *
+from fwi_operators import SourceLike
+
+
+class SourceLikeTTI(SourceLike):
+    def read(self, u, v):
+        eqs = []
+        for i in range(self.npoints):
+            eqs.append(Eq(self[t, i], (self.grid2point(u, self.orig_data[i, :]) + self.grid2point(v, self.orig_data[i, :]))/2))
+        return eqs
+
 
 class TTIOperator(Operator):
     def _init_taylor(self, dim=2, time_order=2, space_order=2):
@@ -18,59 +28,52 @@ class TTIOperator(Operator):
         #
         # with the zero initial conditons to guarantee unicity of the solution
         # Choose dimension (2 or 3)
-
         # half width for indexes, goes from -half to half
-        p=Function('p')
-        r=Function('r')
-        
-        s,h,x,y,z = symbols('s h x y z')
-        m=M(x,y,z)
-        q=Q(x,y,z,t)
-        d=D(x,y,z,t)
-        e=E(x,y,z)
-        
-        A=epsilon(x,y,z)  # (1 + 2epsilon) but make the symbolic simpler
-        B=delta(x,y,z) # sqrt(1 + 2epsilon) but make the symbolic simpler
-        Th=theta(x,y,z)
-        Ph=phi(x,y,z)
+        p = Function('p')
+        r = Function('r')
+
+        s, h, x, y, z = symbols('s h x y z')
+        m = M(x, y, z)
+        # q = Q(x, y, z, t)
+        # d = D(x, y, z, t)
+        e = E(x, y, z)
+
+        A = epsilon(x, y, z)  # (1 + 2epsilon) but make the symbolic simpler
+        B = delta(x, y, z)  # sqrt(1 + 2epsilon) but make the symbolic simpler
+        Th = theta(x, y, z)
+        Ph = phi(x, y, z)
         # Weights to sum the two fields
-        w1=.5
-        w2=.5
-        dttp=as_finite_diff(p(x,y,z,t).diff(t,t), [t-s,t, t+s])
-        dttr=as_finite_diff(r(x,y,z,t).diff(t,t), [t-s,t, t+s])
-        dtp=as_finite_diff(p(x,y,z,t).diff(t), [t-s,t])
-        dtr=as_finite_diff(r(x,y,z,t).diff(t), [t-s,t])
-        # Spacial finite differences can easily be extended to higher order by increasing the list of sampling point in the next expression. 
+        # w1 = .5
+        # w2 = .5
+        dttp = as_finite_diff(p(x, y, z, t).diff(t, t), [t-s, t, t+s])
+        dttr = as_finite_diff(r(x, y, z, t).diff(t, t), [t-s, t, t+s])
+        # dtp = as_finite_diff(p(x, y, z, t).diff(t), [t-s, t])
+        # dtr = as_finite_diff(r(x, y, z, t).diff(t), [t-s, t])
+        # Spacial finite differences can easily be extended to higher order by increasing the list of sampling point in the next expression.
         # Be sure to keep this stencil symmetric and everything else in the notebook will follow.
-        dxxp=as_finite_diff(p(x,y,z,t).diff(x,x), [x-h, x, x+h]) 
-        dyyp=as_finite_diff(p(x,y,z,t).diff(y,y), [y-h, y, y+h]) 
-        dzzp=as_finite_diff(p(x,y,z,t).diff(z,z), [z-h, z, z+h])
-        dxxr=as_finite_diff(r(x,y,z,t).diff(x,x), [x-h, x, x+h]) 
-        dyyr=as_finite_diff(r(x,y,z,t).diff(y,y), [y-h, y, y+h]) 
-        dzzr=as_finite_diff(r(x,y,z,t).diff(z,z), [z-h, z, z+h])
-        
-        
+        dxxp = as_finite_diff(p(x, y, z, t).diff(x, x), [x-h, x, x+h])
+        dyyp = as_finite_diff(p(x, y, z, t).diff(y, y), [y-h, y, y+h])
+        dzzp = as_finite_diff(p(x, y, z, t).diff(z, z), [z-h, z, z+h])
+        dxxr = as_finite_diff(r(x, y, z, t).diff(x, x), [x-h, x, x+h])
+        dyyr = as_finite_diff(r(x, y, z, t).diff(y, y), [y-h, y, y+h])
+        dzzr = as_finite_diff(r(x, y, z, t).diff(z, z), [z-h, z, z+h])
+
         # My 4th order stencil for cross derivatives
-        dxzp = .5/(h**2)*(-2*p(x,y,z,t) + p(x,y,z+h,t) + p(x,y,z-h,t) - p(x+h,y,z-h,t) + p(x-h,y,z,t) - p(x-h,y,z+h,t) + p(x+h,y,z,t))
-        dxzr = .5/(h**2)*(-2*r(x,y,z,t) + r(x,y,z+h,t) + r(x,y,z-h,t) - r(x+h,y,z-h,t) + r(x-h,y,z,t) - r(x-h,y,z+h,t) + r(x+h,y,z,t))
-        
-        dxyp = .5/(h**2)*(-2*p(x,y,z,t) + p(x,y+h,z,t) + p(x,y-h,z,t) - p(x+h,y-h,z,t) + p(x-h,y,z,t) - p(x-h,y+h,z,t) + p(x+h,y,z,t))
-        dxyr = .5/(h**2)*(-2*r(x,y,z,t) + r(x,y+h,z,t) + r(x,y-h,z,t) - r(x+h,y-h,z,t) + r(x-h,y,z,t) - r(x-h,y+h,z,t) + r(x+h,y,z,t))
-        
-        dyzp = .5/(h**2)*(-2*p(x,y,z,t) + p(x,y,z+h,t) + p(x,y,z-h,t) - p(x,y+h,z-h,t) + p(x,y-h,z,t) - p(x,y-h,z+h,t) + p(x,y+h,z,t))
-        dyzr = .5/(h**2)*(-2*r(x,y,z,t) + r(x,y,z+h,t) + r(x,y,z-h,t) - r(x,y+h,z-h,t) + r(x,y-h,z,t) - r(x,y-h,z+h,t) + r(x,y+h,z,t))
-        
+        dxzp = .5/(h**2)*(-2*p(x, y, z, t) + p(x, y, z+h, t) + p(x, y, z-h, t) - p(x+h, y, z-h, t) + p(x-h, y, z, t) - p(x-h, y, z+h, t) + p(x+h, y, z, t))
+        dxzr = .5/(h**2)*(-2*r(x, y, z, t) + r(x, y, z+h, t) + r(x, y, z-h, t) - r(x+h, y, z-h, t) + r(x-h, y, z, t) - r(x-h, y, z+h, t) + r(x+h, y, z, t))
+        dxyp = .5/(h**2)*(-2*p(x, y, z, t) + p(x, y+h, z, t) + p(x, y-h, z, t) - p(x+h, y-h, z, t) + p(x-h, y, z, t) - p(x-h, y+h, z, t) + p(x+h, y, z, t))
+        dxyr = .5/(h**2)*(-2*r(x, y, z, t) + r(x, y+h, z, t) + r(x, y-h, z, t) - r(x+h, y-h, z, t) + r(x-h, y, z, t) - r(x-h, y+h, z, t) + r(x+h, y, z, t))
+        dyzp = .5/(h**2)*(-2*p(x, y, z, t) + p(x, y, z+h, t) + p(x, y, z-h, t) - p(x, y+h, z-h, t) + p(x, y-h, z, t) - p(x, y-h, z+h, t) + p(x, y+h, z, t))
+        dyzr = .5/(h**2)*(-2*r(x, y, z, t) + r(x, y, z+h, t) + r(x, y, z-h, t) - r(x, y+h, z-h, t) + r(x, y-h, z, t) - r(x, y-h, z+h, t) + r(x, y+h, z, t))
         Gxxp = cos(Ph)**2 * cos(Th)**2 * dxxp + sin(Ph)**2 * cos(Th)**2 * dyyp + sin(Th)**2 * dzzp + sin(2*Ph) * cos(Th)**2 * dxyp - sin(Ph) * sin(2*Th) * dyzp - cos(Ph) * sin(2*Th) * dxzp
-        
         Gyyp = sin(Th)**2 * dxxp + cos(Ph)**2 * dyyp - sin(2*Ph)**2 * dxyp
-        
         Gzzr = cos(Ph)**2 * sin(Th)**2 * dxxr + sin(Ph)**2 * sin(Th)**2 * dyyr + cos(Th)**2 * dzzr +\
-        sin(2*Ph) * sin(Th)**2 * dxyr + sin(Ph) * sin(2*Th) * dyzr + cos(Ph) * sin(2*Th) * dxzr
+            sin(2*Ph) * sin(Th)**2 * dxyr + sin(Ph) * sin(2*Th) * dyzr + cos(Ph) * sin(2*Th) * dxzr
         wavep = m * dttp - A * (Gxxp + Gyyp) - B * Gzzr
-        waver = m * dttr - B *  (Gxxp + Gyyp)  - Gzzr 
-        stencilp =solve(wavep,p(x,y,z,t+s),simplify=False)[0]
-        stencilr = solve(waver,r(x,y,z,t+s),simplify=False)[0]
-        return (stencilp, stencilr, (m,A,B,Th,Ph,s, h,e))
+        waver = m * dttr - B * (Gxxp + Gyyp) - Gzzr
+        stencilp = solve(wavep, p(x, y, z, t+s), simplify=False)[0]
+        stencilr = solve(waver, r(x, y, z, t+s), simplify=False)[0]
+        return (stencilp, stencilr, (m, A, B, Th, Ph, s, h, e))
 
     def smart_sympy_replace(self, num_dim, time_order, res, funs, arrs, fw):
         a = Wild('a')
@@ -89,7 +92,7 @@ class TTIOperator(Operator):
                 # Replace function notation with array notation
                 res = res.replace(fun(a, b, c), arr[a, b, c])
                 res = res.replace(arr[a*x+b, c*z+d, e*t+f], arr[e*t+f, a*x+b, c*z+d])
-            else: # num_dim == 3
+            else:  # num_dim == 3
                 res = res.replace(fun(a, b, c, d), arr[a, b, c, d])
                 res = res.replace(arr[x+b, y+q, z+d, t+f], arr[t+f, x+b, y+q, z+d])
         # Replace x+h in indices with x+1
@@ -117,20 +120,14 @@ class TTIOperator(Operator):
 
 
 class ForwardOperator(TTIOperator):
-    def __init__(self, m, src, damp, rec, u, time_order=4, spc_order=12):
+    def __init__(self, m, src, damp, rec, u, v, a, b, th, ph, time_order=4, spc_order=12):
         assert(m.shape == damp.shape)
-        self.input_params = [m, src, damp, rec, u]
+        self.input_params = [m, src, damp, rec, u, v, a, b, th, ph]
         u.pad_time = False
         dt = src.dt
         h = src.h
         old_src = src
-        src = DenseData("src", m.shape, dtype=m.dtype)
-        v = TimeData("v", m.shape, rec.nt, time_order=time_order, save=False, dtype=m.dtype)
-        a = DenseData("a", m.shape, dtype=m.dtype)
-        b = DenseData("b", m.shape, dtype=m.dtype)
-        th = DenseData("th", m.shape, dtype=m.dtype)
-        ph = DenseData("ph", m.shape, dtype=m.dtype)
-        self.output_params = [v, a, b, th, ph]
+        self.output_params = []
         dim = len(m.shape)
         total_dim = self.total_dim(dim)
         space_dim = self.space_dim(dim)
@@ -138,14 +135,11 @@ class ForwardOperator(TTIOperator):
         stencilp = self.smart_sympy_replace(dim, time_order, stencilp, [Function('p'), Function('r')], [u, v], fw=True)
         stencilr = self.smart_sympy_replace(dim, time_order, stencilr, [Function('p'), Function('r')], [u, v], fw=True)
         stencil_args = [m[space_dim], a[space_dim], b[space_dim], th[space_dim], ph[space_dim], dt, h, damp[space_dim]]
-        print(stencil_args)
-        print("****")
-        print(subs)
         first_stencil = Eq(u[total_dim], stencilp)
         second_stencil = Eq(v[total_dim], stencilr)
         self.stencils = [(first_stencil, stencil_args), (second_stencil, stencil_args)]
-        src_list = old_src.add(m, u)
-        rec = rec.read(u)
+        src_list = old_src.add(m, u) + old_src.add(m, v)
+        rec = rec.read(u, v)
         self.time_loop_stencils_post = src_list+rec
         super(ForwardOperator, self).__init__(subs, old_src.nt, m.shape, spc_border=spc_order/2, time_order=time_order, forward=True, dtype=m.dtype)
